more changes for diffusion zone - th
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@506 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/gas_opacity_mod.f90
@@ -308,7 +308,7 @@ subroutine returnGasKappaValue(temperature, rho, lambda, kappaAbs, kappaSca, kap
   type(molecularKappaGrid) :: LookupTable
   real :: temperature
   real, optional :: lambda
-  real :: rho
+  real(double) :: rho
   real, optional :: kappaAbs, kappaSca, kappaAbsArray(:), kappaScaArray(:)
   real :: t1, t2
   real :: pressure, mu
